fix warnings and delete Source.cpp

+++ TheAdTracker/findEyeCorner.cpp
@@ -72,6 +72,6 @@ cv::Point2f findSubpixelEyeCorner(cv::Mat region, cv::Point maxP) {
 	cv::Point maxP2;
 	cv::minMaxLoc(cornerMap, NULL, NULL, NULL, &maxP2);
 
-	return cv::Point2f(sizeRegion.width / 2 + maxP2.x / 10,
-		sizeRegion.height / 2 + maxP2.y / 10);
+	return cv::Point2f(static_cast<float>(sizeRegion.width / 2 + maxP2.x / 10),
+		static_cast<float>(sizeRegion.height / 2 + maxP2.y / 10));
 }
